+++ includes/class-wp-super-duper-widget.php
@@ -62,7 +62,7 @@ class WP_Super_Duper extends WP_Widget {
 	 * Registers the widget with WordPress.
 	 */
 	public function _register() {
-		if ( empty( $this->options['output_types'] ) || in_array( 'widget', $this->options['output_types'] ) ) {
+		if ( ! empty( $this->options['output_types'] ) && in_array( 'widget', $this->options['output_types'], true ) ) {
 			parent::_register();
 		}
 	}